refactor: updating imports & implementing new hooks
