+++ src/Pages/Orders/index.tsx
@@ -4,9 +4,9 @@ import { CustomTh } from '../../components/CustomTh';
 import { PlusCircle } from 'lucide-react';
 import { IconButton } from '../../components/IconButton';
 import { useNavigate } from 'react-router';
-import { OrderEntity } from '../../constants/order';
+import { OrderEntity } from '../../types/order';
 import { PageContainer } from '../../components/PageContainer';
-import { EmployeeEntity } from '../../constants/employee';
+import { EmployeeEntity } from '../../types/employee';
 import { toFullLocaleDate } from '../../utils/toFullLocaleDate';
 import { Controller, useForm } from 'react-hook-form';
 import { useEffect, useRef, useState } from 'react';
