--- src/constants/product.ts
@@ -1,10 +1,0 @@
-export type ProductEntity = {
-    id: string;
-    name: string;
-    code: number;
-    value: number;
-    category: string;
-    stock: number;
-    active: boolean;
-};
-
